not generics, but seperate lra configuration
